+++ cmd/analyze.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/dpcat237/go-dsu/internal/analyzer"
+
 	"github.com/spf13/cobra"
 
 	"github.com/dpcat237/go-dsu/internal/download"
@@ -12,28 +14,27 @@ import (
 	"github.com/dpcat237/go-dsu/internal/logger"
 	"github.com/dpcat237/go-dsu/internal/module"
 	"github.com/dpcat237/go-dsu/internal/output"
-	"github.com/dpcat237/go-dsu/internal/previewer"
 	"github.com/dpcat237/go-dsu/internal/vulnerability"
 )
 
 var (
-	previewCmd = &cobra.Command{
-		Use:   "preview",
-		Short: "Preview updates",
-		Long:  `Preview available updates of direct modules with changes`,
+	analyzeCmd = &cobra.Command{
+		Use:   "analyze",
+		Short: "Analyze current state",
+		Long:  `Analyze licenses and vulnerabilities of current dependencies`,
 		Run: func(cmd *cobra.Command, args []string) {
-			preview(cmd)
+			analyze(cmd)
 		},
 	}
 )
 
 func init() {
-	rootCmd.AddCommand(previewCmd)
-	previewCmd.Flags().Bool("dev", false, "Development mode")
-	previewCmd.Flags().String("path", "", "Preview project from git path. Eg. github.com/spf13/cobra")
+	rootCmd.AddCommand(analyzeCmd)
+	analyzeCmd.Flags().Bool("dev", false, "Development mode")
+	analyzeCmd.Flags().String("path", "", "Preview project from git path. Eg. github.com/spf13/cobra")
 }
 
-func preview(cmd *cobra.Command) {
+func analyze(cmd *cobra.Command) {
 	mod := output.ModeProd
 	if cmd.Flag("dev").Value.String() == "true" {
 		mod = output.ModeDev
@@ -61,7 +62,7 @@ func preview(cmd *cobra.Command) {
 	dwnHnd := download.InitHandler(exc, lgr)
 	vlnHnd := vulnerability.InitHandler(lgr)
 	hnd := module.InitHandler(exc)
-	prw := previewer.Init(dwnHnd, exc, lgr, licHnd, hnd, vlnHnd)
-	out := prw.Preview(cmd.Flag("path").Value.String())
+	anz := analyzer.Init(dwnHnd, exc, lgr, licHnd, hnd, vlnHnd)
+	out := anz.AnalyzeDependencies(cmd.Flag("path").Value.String())
 	fmt.Println(out.ToString(mod))
 }